glib: Provide `ThreadGuard` and `thread_id()` as public API
It's useful in other contexts than GLib too.
